Completed surveillance module refactor
