+++ controller/surveillanceCtrl.js
@@ -1,0 +1,25 @@
+import * as SurveillanceDB from '../db/record-surveillance.js';
+
+// Get surveillance Record Page
+export const getSurveillanceRecordPage = async (req, res) => {
+  const recordSurveillanceData = await SurveillanceDB.getAllRecordSurveillance();
+
+  res.status(200)
+    .render('surveillance-record', { recordSurveillanceData });
+};
+
+// Update Surveillance Status
+export const updateSurveillanceRecord = async (req, res) => {
+  try {
+    const id = req.query.id;
+    const result = await SurveillanceDB.updateSurveillanceStatus(id);
+    if (result) {
+      res.status(200)
+        .redirect('/home');
+    }
+  } catch (error) {
+    console.error('Error during updating coop record', error);
+    res.status(500)
+      .send('Internal Server Error');
+  }
+};
