+++ views/incubator-record.ejs
@@ -15,7 +15,7 @@
 <% 
   
   // Button disable for Monday & Thursday
-  let insertTrayFlag = "";
+  let insertTrayFlag = "disabled";
   const currDate = new Date(); 
   const currDay = currDate.getDay(); // Return 0 for Sunday, 1 for Monday
   if (currDay === 1 || currDay === 4) {
@@ -23,8 +23,8 @@
   }
   
   // Button disable for hatching date
-  let inc1_flag = "";
-  let inc2_flag = "";
+  let inc1_flag = "disabled";
+  let inc2_flag = "disabled";
   if (!hatchingDate || hatchingDate.length == 0) {
     console.log("No hatching today")
   } else {
@@ -128,7 +128,7 @@
           </a>
         </li>
         <li class="nav-item">
-          <a class="nav-link text-white " href="/surveillance-record">
+          <a class="nav-link text-white " href="/surveillance/view">
             <div class="text-white text-center me-2 d-flex align-items-center justify-content-center">
               <i class="material-icons opacity-10">visibility</i>
             </div>
